+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## 2025.10.09
+
+### Added
+
+- MusicDatabase1.java, which includes internal class Song, a comparator for comparing the title of songs alphabetically, a constructor for musicDatabase1, versions of addEntry, and getEntry. It also includes a main method used to demonstrate the methods.
+
 ## 2025.09.21
 
 ### Added